\name{HWEintrinsic-package}
\alias{HWEintrinsic-package}
\alias{HWEintrinsic}
\docType{package}
\title{
\packageTitle{HWEintrinsic}
}
\description{
\packageDescription{HWEintrinsic}
}
\details{
  The package is loaded with the usual \code{library(HWEintrinsic)} command. The most important functions are \code{\link{hwe.ibf}}, \code{\link{hwe.ibf.mc}}, and \code{\link{hwe.ibf.plot}}.
  It contains also some data sets which have been extensively used in the literature.   
}
\author{
\packageAuthor{HWEintrinsic}

Maintainer: \packageMaintainer{HWEintrinsic}
}
\references{
  Consonni, G., Moreno, E., and Venturini, S. (2011). "Testing Hardy-Weinberg equilibrium: an objective Bayesian analysis". Statistics in Medicine, \bold{30}, 62--74. \url{https://onlinelibrary.wiley.com/doi/10.1002/sim.4084/abstract}
}
\seealso{
   \code{\link{hwe.ibf}}, 
   \code{\link{hwe.ibf.mc}},
   \code{\link{hwe.ibf.plot}},
   \code{\link{hwe.bf}}.
}
\keyword{intrinsic prior}
\keyword{Bayes Factor}
\keyword{Hardy-Weinberg}
\keyword{HWE}
